requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
defusedxml==0.6.0
Django==2.2.4
django-cors-headers==3.1.0
django-filter==2.2.0
django-otp==0.6.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
drf-yasg==1.16.1
ecdsa==0.13.2
future==0.17.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==19.1
psycopg2==2.8.3
pyasn1==0.4.6
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.2
python-dotenv==0.10.3
python-jose==3.0.1
python3-openid==3.1.0
pytz==2019.2
qrcode==6.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
